lib: use 32-bit (not 64-bit) sector addresses

+++ lib/dev.h
@@ -12,15 +12,15 @@ struct jgfs2_dev {
 	int         fd;
 	
 	uint64_t size_byte;
-	uint64_t size_sect;
+	uint32_t size_sect;
 };
 
 
 extern struct jgfs2_dev dev;
 
 
-void *jgfs2_dev_map_sect(uint64_t sect_num, uint64_t sect_count);
-void jgfs2_dev_unmap_sect(void *addr, uint64_t sect_num, uint64_t sect_count);
+void *jgfs2_dev_map_sect(uint32_t sect_num, uint32_t sect_count);
+void jgfs2_dev_unmap_sect(void *addr, uint32_t sect_num, uint32_t sect_count);
 
 void jgfs2_dev_fsync(void);
 void jgfs2_dev_msync(void *addr, size_t length);
